src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "sqlite-vec>=0.1.6",
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.137.0",
    "Pillow>=10.3.0",
    "google-generativeai>=0.7.2",
]
